move class definition into a new file
